Some refactoring
- COMPOSE_MAPPING variable autoloading based on cli options
  upon dispatch module load removed. Now the code that needs
  access to COMPOSE_MAPPING will load it on-demand via
  dispatch.get_compose_mapping method.
- dispatch global variables have been moved out to dispatch_globals.py

OAMG-9529
