+++ src/dispatch/__main__.py
@@ -5,17 +5,17 @@ import sys
 
 import requests
 
-from dispatch.__init__ import (
-    ARTIFACT_MAPPING,
-    COMPOSE_MAPPING,
-    PACKAGE_MAPPING,
+from dispatch import (
     get_arguments,
+    get_compose_mapping,
     get_logging,
 )
+from dispatch.dispatch_globals import ARTIFACT_MAPPING, PACKAGE_MAPPING
 from dispatch.tf_send_request import submit_test
 
 LOGGER = get_logging()
 ARGS = get_arguments()
+COMPOSE_MAPPING = get_compose_mapping()
 
 
 def main():
