+++ src/dispatch/brew_api.py
@@ -3,13 +3,15 @@ import sys
 
 import koji
 
-from dispatch.__init__ import COMPOSE_MAPPING, get_arguments, get_logging
+from dispatch import get_compose_mapping, get_arguments, get_logging
 
 LOGGER = get_logging()
 ARGS = get_arguments()
 
 SESSION = koji.ClientSession("https://brewhub.engineering.redhat.com/brewhub")
 SESSION.gssapi_login()
+
+COMPOSE_MAPPING = get_compose_mapping()
 
 EPEL_COMPOSES = {
     "rhel-8": [
